+++ bolt/reports/sql/view_mb_deadheads.sql
@@ -1,0 +1,10 @@
+CREATE OR REPLACE VIEW view_mb_deadheads AS SELECT
+    YMTH,
+    Service,
+    ROUND(SUM("Pull-In/Out Distance (A)"), 2) AS "Deadhead Miles",
+    ROUND(SUM("Pull-In/Out Hours (A)"), 2) AS "Deadhead Hours"
+FROM CR0174
+GROUP BY
+    YMTH,
+    Service
+;