+++ tasks.py
@@ -5,6 +5,7 @@ For now... might migrate to click.
 import datetime as dt
 import json
 from hashlib import sha256
+from pathlib import Path
 
 from invoke import task
 from rich.console import Console
@@ -129,6 +130,33 @@ def update(c, datasource: str="all", cache: int=1):  # TODO: flag to cache data
         console.print(f"- [blue]{name}[/]: [red]{err}[/]")
     return
 
+
+# ============================================================================
+# WIP
+
+@task
+def status(c):  # TODO: WIP
+    """A git-like status command that lists unregistered datasets."""
+    console.print("Data files not registered:")
+    console.print('(use "invoke add <file>..." to register new files)')
+    for i in [
+        r"C:\Workspace\tmpdb\Data\raw\CR - CR0174\202301-Weekday-CR-0174 Incident Adjusted Distance and Time - Jan 13 2025.csv",
+        r"C:\Workspace\tmpdb\Data\raw\CR - CR0174\202301-Saturday-CR-0174 Incident Adjusted Distance and Time - Jan 13 2025.csv",
+        r"C:\Workspace\tmpdb\Data\raw\CR - CR0174\202301-Sunday-CR-0174 Incident Adjusted Distance and Time - Jan 13 2025.csv",
+    ]:
+        console.print(f"        [red]added:    {Path(i).name}[/]")
+
+@task
+def add(c):
+    """A git-like add command that adds unregistered datasets."""
+    ...
+
+
+@task
+def update_db(c):
+    """PLACEHOLDER"""
+    bolt.datasources.warehouse.load_cache_files(compact_db=True)
+    return
 
 # ============================================================================
 # REPORT RUNNERS